requirements.txt
# Core dependencies
fastapi>=0.95.0
uvicorn>=0.17.0
pydantic>=1.9.0,<2.0.0
aiohttp>=3.8.1
websockets>=10.3
python-multipart>=0.0.5

# Security
pyjwt>=2.3.0
cryptography>=36.0.0
passlib>=1.7.4

# RAG system
chromadb>=0.4.6
sentence-transformers>=2.2.2
numpy>=1.22.3

# Home Assistant
homeassistant-api>=4.1.0

# Utilities
python-dotenv>=0.20.0
requests>=2.27.1
jinja2>=3.0.3
pyyaml>=6.0
tqdm>=4.64.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.18.3
pytest-cov>=3.0.0